+++ CHANGELOG.md
@@ -12,11 +12,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Frost Trees will now drop saplings
 - Saplings are now pottable
 - Aspect slabs can now be used in crafting recipes
+- Uranium button now made with 1 raw uranium instead of 1 uranium block
+- Removed default name visibility of Lotus Flower
 
 ### Fixed
 
 - Fix player gutter multiplier, which wasn't being saved
 - Space sapling texture baseline lowered, space and frost sapling sides trimmed
+- Added null check for target in MobAnimationPhaseGoal
 
 ### Contributors for this release